requirements.txt
absl-py==2.1.0
accelerate==0.34.2
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
asttokens==3.0.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.45.0
boltons==24.1.0
Brotli==1.1.0
certifi==2022.12.7
cffi==1.17.1
chardet==5.2.0
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
DataProperty==1.0.1
datasets==3.0.0
decorator==5.1.1
deepspeed==0.15.1
dill==0.3.8
distro==1.9.0
dnspython==2.4.2
docker-pycreds==0.4.0
einops==0.8.0
et-xmlfile==1.1.0
evaluate==0.4.3
exceptiongroup==1.2.2
executing==2.1.0
expecttest==0.1.6
fastapi==0.115.6
filelock==3.13.1
fire==0.6.0
flash-attn==2.6.3
frozenlist==1.4.1
fschat==0.2.36
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.43
gmpy2==2.2.1
h11==0.14.0
hjson==3.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.25.0
hypothesis==6.88.4
idna==3.4
intel-cmplr-lib-ur==2025.0.2
intel-openmp==2025.0.2
ipython==8.30.0
jedi==0.19.2
Jinja2==3.1.3
jiter==0.8.0
joblib==1.4.2
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==2.1
latex2mathml==3.77.0
libarchive-c==5.1
lxml==5.3.0
markdown-it-py==3.0.0
markdown2==2.5.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mbstrdecoder==1.1.3
mdurl==0.1.2
mkl==2025.0.1
mkl-service==2.4.0
mkl_fft==1.3.11
mkl_random==1.2.8
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.2.1
nh3==0.2.19
ninja==1.11.1.1
nltk==3.9.1
numexpr==2.10.1
numpy==1.26.4
nvidia-ml-py==12.560.30
openai==0.28.0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.2
parso==0.8.4
pathvalidate==3.2.1
peft==0.12.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pkginfo==1.11.2
platformdirs==4.3.6
pluggy==1.5.0
portalocker==2.10.1
prompt_toolkit==3.0.48
protobuf==5.28.2
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==17.0.0
pybind11==2.13.6
pycosat==0.6.6
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyOpenSSL==24.3.0
PySocks==1.7.1
pytablewriter==1.2.0
python-dateutil==2.9.0.post0
python-etcd==0.4.5
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rich==13.9.4
rouge_score==0.1.2
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
sacrebleu==2.4.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentencepiece==0.2.0
sentry-sdk==2.19.2
setproctitle==1.3.4
shellingham==1.5.4
shortuuid==1.0.13
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlitedict==2.1.0
stack-data==0.6.3
starlette==0.41.3
svgwrite==1.4.3
sympy==1.13.1
tabledata==1.3.3
tabulate==0.9.0
tbb==2022.0.0
tcmlib==1.2.0
tcolorpy==0.1.6
tensorboardX==2.6.2.2
termcolor==2.4.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.19.1
tomli==2.2.1
toolz==1.0.0
torchelastic==0.2.2
tqdm==4.66.5
tqdm-multiprocess==0.0.11
traitlets==5.14.3
transformers==4.40.1
triton==2.1.0
truststore==0.10.0
typepy==1.3.2
typer==0.15.1
types-dataclasses==0.6.6
typing_extensions==4.12.2
tzdata==2024.1
umf==0.9.1
urllib3==1.26.13
uvicorn==0.32.1
wandb==0.19.0
wavedrom==2.0.3.post3
wcwidth==0.2.13
word2number==1.1
xxhash==3.5.0
yarl==1.11.1
zstandard==0.23.0
